requirements.txt
Django==2.0.6
djangorestframework==3.8.2
Pillow==5.1.0
pytz==2018.5
